requirements.txt
# 동전게임 실시간 공략 가이드 - 의존성 패키지

# AI/Vision
google-generativeai>=0.3.0

# 이미지 처리
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# 화면 캡처
mss>=9.0.0
pygetwindow>=0.0.9

# 물리 엔진
pymunk>=6.5.0

# GUI (tkinter는 Python 기본 포함)
# tkinter는 별도 설치 불필요 (Python 표준 라이브러리)
